front: fix and refactor selected train id init
