+++ front/src/applications/operationalStudies/views/Scenario/ScenarioView.tsx
@@ -2,7 +2,6 @@ import { useState } from 'react';
 
 import useScenario from 'applications/operationalStudies/hooks/useScenario';
 import { ScenarioContextProvider } from 'applications/operationalStudies/hooks/useScenarioContext';
-import useScenarioQueryParams from 'applications/operationalStudies/hooks/useScenarioQueryParams';
 import { RollingStockContextProvider } from 'common/RollingStockContext';
 import { SubCategoryContextProvider } from 'common/SubCategoryContext';
 import useInfraStatus from 'modules/pathfinding/hooks/useInfraStatus';
@@ -14,9 +13,6 @@ import ScenarioContent from './components/ScenarioContent';
 const Scenario = () => {
   const { scenario } = useScenario();
   const { infra } = useInfraStatus({ infraId: scenario?.infra_id });
-
-  // Initialize and sync the URL and local storage with Redux
-  useScenarioQueryParams();
 
   const [activeBoards, setActiveBoards] = useState<Set<Board>>(
     new Set<Board>(['trains', 'map', 'std', 'sdd', 'tables'])
